requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
scikit-learn
nltk==3.8.1
numpy
pandas
joblib==1.3.1
python-dotenv==1.0.0
gunicorn
lxml
urllib3==2.0.4
